This is a rather large commit, also serving as a comparison of how to change your commit messages / bring multiple commits together.  Note that, in suit with what we talked about in lecture, this is really only appropriate because I am the only person working on this repository.
If I only wanted to change the commit messages, I could use "git commit --amend".  This option is particularly useful if
you have not yet pushed your commit, but is also good for clarifying a commit if you need to!

In lecture, I added Lecture09's changes as well as lectures 10 and 11 in one commit, call it commit A.

Right after, I added / committed lecture 16.

Neither of the commit messages were useful in terms of what the contents are, and I am going to modify these to include a
better message as well as demonstrate the difference between a "soft" reset and a "hard" reset.  The goal here is to bring
the two commits together into one, and provide a more descriptive commit message all at once.

The command I executed to get back in time:

>>> git reset --soft HEAD~2

This took me back two commits, but since I did --soft, when I check the status I will see what the candidates for committing
are, with some annotations in the status for emphasis:

                                    two commits
>>> git status                          VVV
On branch master                      ############             ################
Your branch is behind 'origin/master' by 2 commits, and can be fast-forwarded.## No merge conflicts (obviously)!
  (use "git pull" to update your local branch)                 ################
Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

    modified:   lecture09/lecture09.pdf
    modified:   lecture09/lecture09.xetex
    new file:   lecture10/lecture10.pdf
    new file:   lecture10/lecture10.xetex
    new file:   lecture10/references.bib
    new file:   lecture11/lecture11.pdf
    new file:   lecture11/lecture11.xetex
    new file:   lecture11/references.bib
    new file:   lecture16/lecture16.pdf   #
    new file:   lecture16/lecture16.xetex #### This was the second commit, but I have them all staged for change now.
    new file:   lecture16/references.bib  #
    new file:   lecture16/tmux.jpg        #

Lecture09:
  - Added an example of the "until" loop to the end of the lecture.

Lecture10:
  - This lecture covered the differences between the "case" and "if" statements in bash, and was largely comprised
    of in-class examples.
  - I have included clearer explanations of how to use both, as well as an explanation of the default behavior of
    the comparison operations (which differ between case and if).
  - I decided to include how to wield the =~ binary operator in if statements in bash by comparing with extended
    regular expressions, as well as some examples of how to use these structures.
  - I filled in the slides with the "words" to accompany explaining what merging in git is, when it happens, and
    how to configure your "git difftool" in order to make comparisons significantly more convenient.

Lecture11:
  - This lecture was a walkthrough of many of the different (and fascinating!) things you can do with bash arrays.
  - I have designed the slides so that even if you missed lecture, you should be able to follow along copy-pasting
    from the slides to get a roughly equivalent experience.
  - For those who were in lecture, I explained some of the concepts we covered with words which may be a good refresher.
  - The link in [2] in the bibliography has a wealth of complex array capabilities that are definitely worth seeing.

Lecture16:
  - This is what was presented to you during lecture, the only change being that I showed how to put your git tags
    online at the end.

Note that using this strategy I will need to "git push -f", just as we saw with "git reset --hard".  The primary
difference for you to see here is that with a "hard" reset, the changes do not remain staged to commit (will not
show up with "git status").

There is a lot more that can be done with changing history, an excellent resource is available here:

https://git-scm.com/book/en/v2/Git-Tools-Rewriting-History

As I emphasized in lecture, please be very careful when using the "git reset" command, ESPECIALLY when you are
working on a branch / repository with other people!  You may cause some serious grief that would have better been
resolved with a "git revert" instead, to preserve the history / avoid "lost-in-time" conflicts.